Moved Copyable to consensus-utility
