+++ platform-sdk/consensus-utility/src/main/java/org/hiero/consensus/utility/Copyable.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common;
+package org.hiero.consensus.utility;
 
 /**
  * An object that can be copied.
