+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/FastCopyable.java
@@ -3,6 +3,7 @@ package com.swirlds.common;
 
 import com.swirlds.base.state.Mutable;
 import org.hiero.consensus.model.crypto.Hashable;
+import org.hiero.consensus.utility.Copyable;
 
 /**
  * An interface for classes that can be copied and serialized in a way specific to the Swirlds platform. If a class
